Create DeferredResolvableCollectionTrait
